Added a validation mixin
