+++ app/controllers/api/v1x0/mixins/validation_mixin.rb
@@ -1,0 +1,15 @@
+module Api
+  module V1x0
+    module Mixins
+      module ValidationMixin
+        def group_id_array_check(uuids)
+          if !uuids.kind_of?(Array)
+            invalid_parameter('Group should be an array')
+          elsif uuids.blank? || uuids.any?(&:blank?)
+            invalid_parameter('Group should not be empty')
+          end
+        end
+      end
+    end
+  end
+end
